requirements.txt
django
pillow
django-htmx
whitenoise
PyMySQL
python-dotenv